Add popup open heartbeat, refactor background.ts (#1131)
