+++ apps/browser-extension/src/i18n/locales/en.json
@@ -324,7 +324,7 @@
     "autoLockTimeoutDescription": "Automatically lock the vault after a period of inactivity",
     "autoLockTimeoutHelp": "The timer resets when you open the browser extension or use autofill on a webpage. The vault will only lock after the specified period of inactivity (no autofill usage or extension popup opened). The vault will always lock when the browser is closed, regardless of this setting.",
     "autoLockNever": "Never",
-    "autoLock5Seconds": "5 seconds (debug)",
+    "autoLock15Seconds": "15 seconds",
     "autoLock1Minute": "1 minute",
     "autoLock5Minutes": "5 minutes",
     "autoLock15Minutes": "15 minutes",
